restructured components and pages
