+++ src/App.tsx
@@ -1,7 +1,6 @@
 import * as React from "react";
 import { Switch } from "react-router-dom";
-import { Footer } from "./components/Footer";
-import { Navbar } from "./components/Navbar";
+import { Navbar, Footer } from "./components/exportComponents";
 import { Routes } from "./utils/Routes";
 
 export const App = () => (
