+++ src/components/Layout/Navbar.tsx
@@ -2,8 +2,8 @@
 import { useApolloClient } from '@apollo/client';
 import { Box, Button, Flex, Link, useColorMode, useToast } from "@chakra-ui/react";
 import React from 'react';
-import { ColorModeSwitcher } from '../ColorModeSwitcher';
-import { useLogoutMutation, useMeQuery } from "../generated/graphql";
+import { ColorModeSwitcher } from '../../ColorModeSwitcher';
+import { useLogoutMutation, useMeQuery } from "../../generated/graphql";
 
 interface NavbarProps {
