+++ src/pages/authentication/register.tsx
@@ -3,10 +3,9 @@ import { Box, Heading, Link, useColorModeValue, Text } from '@chakra-ui/react';
 import { Form, Formik } from 'formik';
 import React from 'react';
 import { RouteComponentProps } from 'react-router';
-import { InputField } from '../components/InputField';
-import { Responsive } from '../components/Responsive';
-import { MeDocument, MeQuery, useRegisterMutation } from '../generated/graphql';
-import { toErrorMap } from '../utils/toErrorMap';
+import { Responsive, InputField } from '../../components/exportComponents';
+import { MeDocument, MeQuery, useRegisterMutation } from '../../generated/graphql';
+import { toErrorMap } from '../../utils/toErrorMap';
 
 
 export const Register: React.FC<RouteComponentProps> = ({ history }) => {
