+++ src/pages/create/page/create-page.tsx
@@ -3,17 +3,16 @@ import { Button, Image } from '@chakra-ui/react';
 import { Form, Formik } from 'formik';
 import React, { useRef, useState } from 'react';
 import { RouteComponentProps } from 'react-router';
+import { useCreatePageMutation } from '../../../generated/graphql';
+import { cloudinarySignature } from '../../../utils/utilCloudinary';
 import { v4 as uuidv4 } from 'uuid';
-import { InputField } from '../components/InputField';
-import { Responsive } from '../components/Responsive';
-import { useCreatePostMutation } from '../generated/graphql';
-import { cloudinarySignature } from '../utils/utilCloudinary';
+import { Responsive, InputField } from '../../../components/exportComponents';
 
-export const CreatePost: React.FC<RouteComponentProps> = ({ history }) => {
+export const CreatePage: React.FC<RouteComponentProps> = ({ history }) => {
     const [file, setFile] = useState<File>();
     const [fileUrl, setFileUrl] = useState<string>();
     const fileInputRef = useRef<HTMLInputElement>(null);
-    const [createPost] = useCreatePostMutation();
+    const [createPage] = useCreatePageMutation();
 
     const uploadImage = async () => {
         if (!file) {
@@ -67,48 +66,70 @@ export const CreatePost: React.FC<RouteComponentProps> = ({ history }) => {
         setFileUrl(URL.createObjectURL(newFile));
     }
 
+    let body = null;
     return (
         <Responsive variant="small">
             <Formik
-                initialValues={{ title: '', text: '' }}
+                enableReinitialize
+                initialValues={{ pageTitle: '', pageText: '', aboutUs: '' }}
                 onSubmit={async (values, { setErrors, resetForm }) => {
                     let imgUrl: string | undefined;
 
                     if (file) {
                         const upload = await uploadImage();
-                        if (!upload.success) return;
+                        if (!upload.success) {
+                            //if upload is not successful
+                            return; //return nothing
+                        }
 
+                        //if successful 
                         imgUrl = upload.url;
+                        console.log("formik this is the img url", imgUrl);
                     }
 
-                    const response = await createPost({ variables: { postimgUrl: imgUrl, input: values } });
+                    const response = await createPage({
+                        variables: { ...values, pageimgUrl: imgUrl }
+                    })
+                    //if there is an error
+                    //TO DO
+                    // Fix the error message
                     if (response.errors) {
-                        setErrors({ title: 'an error occured' });
+                        setErrors({ pageTitle: "this is an error", pageText: "error on text" });
                         return;
                     }
                     resetForm();
-                    const post = response.data?.createPost.postCreatorId;
-                    history.push(`/pages/${post}`);
+                    const page = response.data?.createPage.id;
+                    console.log("page is posted: ", page);
+                    history.push("/");
                 }}
+
             >
                 {({ isSubmitting }) => (
-                    <Form id="newpost" style={{ width: "100%" }}>
+                    <Form>
                         <InputField
-                            name="title"
+                            name="pageTitle"
                             placeholder="pageTitle"
                             label="Title"
                         />
                         <Box mt={4}>
                             <InputField
                                 textarea
-                                name="text"
+                                name="pageText"
                                 placeholder="text..."
                                 label="Description"
                             />
                         </Box>
                         <Box mt={4}>
+                            <InputField
+                                textarea
+                                name="aboutUs"
+                                placeholder="text..."
+                                label="About us"
+                            />
+                        </Box>
+                        <Box mt={4}>
                             <Button ml={12} type="submit" isLoading={isSubmitting}>
-                                Create Post
+                                Create Page
                             </Button>
                             <label htmlFor="postImage">
                                 <Button mr={8} onClick={() => fileInputRef.current?.click()}>
@@ -128,6 +149,7 @@ export const CreatePost: React.FC<RouteComponentProps> = ({ history }) => {
                     </Form>
                 )}
             </Formik>
+            {body ? body : null}
         </Responsive>
     );
 }
