+++ src/pages/exportPage.ts
@@ -1,0 +1,13 @@
+export * from './authentication/forgot-password';
+export * from './authentication/login';
+export * from './authentication/register';
+export * from './changePassword/[token]';
+export * from './create/page/[id]';
+export * from './create/page/create-page';
+export * from './create/page/editPage/[id]';
+export * from './create/post/create-post';
+export * from './create/post/editPost/[id]';
+export * from './landing/home';
+
+
+
